finishes up moving code into handlers
